Applied Magnus Hagander's patch to take away some compiler warnings.

+++ src/interfaces/ecpg/ecpglib/typename.c
@@ -96,6 +96,6 @@ ECPGDynamicType(Oid type)
 		case NUMERICOID:
 			return SQL3_NUMERIC;	/* numeric */
 		default:
-			return -type;
+			return -(int)type;
 	}
 }
